Fix to_int32

+++ hanser/transform/segmentation.py
@@ -119,7 +119,7 @@ def _crop(image, offset_height, offset_width, crop_height, crop_width):
             tf.greater_equal(original_shape[1], crop_width)),
         ['Crop size greater than the image size.'])
 
-    offsets = tf.to_int32(tf.stack([offset_height, offset_width, 0]))
+    offsets = tf.cast(tf.stack([offset_height, offset_width, 0]), tf.int32)
 
     # Use tf.slice instead of crop_to_bounding box as it accepts tensors to
     # define the crop size.
